refactor(classification): split classification_classes into datasets subpackage
Extracts each classification system (USCS, EN main, lithology, color) into its own module under utils/datasets/, with the abstract base class and LayerInformation in classification.py and ExistingClassificationSystems registry in __init__.py. Updates all imports across classifiers, models, evaluation, and API endpoints.
