+++ src/classification/classifiers/classifier_factory.py
@@ -8,7 +8,7 @@ from classification.classifiers.baseline_classifier import BaselineClassifier
 from classification.classifiers.bert_classifier import BertClassifier
 from classification.classifiers.classifier import ClassifierTypes
 from classification.classifiers.dummy_classifier import DummyClassifier
-from classification.utils.classification_classes import ClassificationSystem
+from classification.utils.datasets.classification import ClassificationSystem
 
 logger = logging.getLogger(__name__)
