+++ src/classification/utils/data_utils.py
@@ -11,7 +11,7 @@ from dataclasses import dataclass
 from pathlib import Path
 
 from classification.evaluation.evaluate import AllClassificationMetrics
-from classification.utils.classification_classes import ClassificationSystem, LayerInformation
+from classification.utils.datasets.classification import ClassificationSystem, LayerInformation
 from classification.utils.file_utils import read_params
 
 classification_params = read_params("classification_params.yml")
