+++ src/classification/utils/datasets/__init__.py
@@ -1,0 +1,37 @@
+"""TODO."""
+
+from enum import Enum
+from typing import Literal
+
+from classification.utils.datasets.classification import ClassificationSystem
+from classification.utils.datasets.color import ColorConsolidatedSystem
+from classification.utils.datasets.en_main import ENMainSystem
+from classification.utils.datasets.lithology import LithologySystem
+from classification.utils.datasets.uscs import USCSSystem
+
+
+class ExistingClassificationSystems(Enum):
+    """Enum listing all existing classification types.
+
+    The value of each entry is the Classification system class, not an instance of the class.
+    """
+
+    uscs = USCSSystem
+    lithology = LithologySystem
+    en_main = ENMainSystem
+    color = ColorConsolidatedSystem
+
+    @classmethod
+    def get_classification_system_type(
+        cls, class_system: Literal["uscs", "lithology", "en_main", "color"]
+    ) -> type[ClassificationSystem]:
+        """Returns the class of a classification system based on input string.
+
+        Args:
+            class_system (Literal["uscs", "lithology", "en_main"]): The name of the classification system.
+
+        Returns:
+            Type[ClassificationSystem]: The associated ClassificationSystem class.
+
+        """
+        return cls[class_system].value
